Add files via upload
This is a very early version of an update script that I put together to update my Fedora system. There is extra stuff in here that is only there for a visual aspect and not necessary in completing the purpose of the script. This will likely change as I learn more and evolve this script over time.