Update setup.sh
Refining some things in this setup script. I am aware that I can list all pieces of software in a single command but I like it laid out this way. It makes it easier for me to read.

Still working on the software section. Going through what I have on my other computer as I put this script together.